Refactor Post management by restructuring resources
- Moved PostResource to a new namespace: Firefly\FilamentBlog\Resources\Posts.
- Introduced new schemas for PostForm and PostInfolist to streamline form and infolist configurations.
- Added new pages for creating, editing, listing, and viewing posts.
- Removed deprecated methods and classes related to post management to enhance clarity and maintainability.
- Updated Blog.php to reflect the new PostResource path.
