--- src/Resources/PostResource/RelationManagers/SeoDetailRelationManager.php
@@ -1,52 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources\PostResource\RelationManagers;
-
-use Filament\Schemas\Schema;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Actions\CreateAction;
-use Filament\Actions\EditAction;
-use Filament\Actions\DeleteAction;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Filament\Resources\RelationManagers\RelationManager;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Firefly\FilamentBlog\Models\SeoDetail;
-
-class SeoDetailRelationManager extends RelationManager
-{
-    protected static string $relationship = 'seoDetail';
-
-    public function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components(SeoDetail::getForm());
-    }
-
-    public function table(Table $table): Table
-    {
-        return $table
-            ->recordTitleAttribute('title')
-            ->columns([
-                TextColumn::make('title'),
-                TextColumn::make('description'),
-                TextColumn::make('keywords')->badge(),
-            ])
-            ->filters([
-                //
-            ])
-            ->headerActions([
-                CreateAction::make(),
-            ])
-            ->recordActions([
-                EditAction::make()->slideOver(),
-                DeleteAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-}
