+++ src/Resources/Posts/Schemas/PostInfolist.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Firefly\FilamentBlog\Resources\Posts\Schemas;
+
+use Filament\Infolists\Components\TextEntry;
+use Filament\Schemas\Components\Fieldset;
+use Filament\Schemas\Components\Section;
+use Filament\Schemas\Schema;
+use Firefly\FilamentBlog\Enums\PostStatus;
+use Firefly\FilamentBlog\Models\Post;
+
+class PostInfolist
+{
+    public static function configure(Schema $schema): Schema
+    {
+        return $schema->components([
+            Section::make('Post')
+                ->columnSpanFull()
+                ->schema([
+                    Fieldset::make('General')
+                        ->schema([
+                            TextEntry::make('title'),
+                            TextEntry::make('slug'),
+                            TextEntry::make('sub_title'),
+                        ]),
+                    Fieldset::make('Publish Information')
+                        ->schema([
+                            TextEntry::make('status')
+                                ->badge()->color(function ($state) {
+                                    return $state->getColor();
+                                }),
+                            TextEntry::make('published_at')->visible(function (Post $record) {
+                                return $record->status === PostStatus::PUBLISHED;
+                            }),
+
+                            TextEntry::make('scheduled_for')->visible(function (Post $record) {
+                                return $record->status === PostStatus::SCHEDULED;
+                            }),
+                        ]),
+                    Fieldset::make('Description')
+                        ->schema([
+                            TextEntry::make('body')
+                                ->html()
+                                ->columnSpanFull(),
+                        ]),
+                ]),
+        ]);
+    }
+}
