+++ src/Resources/Posts/Tables/PostsTable.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Firefly\FilamentBlog\Resources\Posts\Tables;
+
+use Filament\Actions\ActionGroup;
+use Filament\Actions\BulkActionGroup;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Actions\ViewAction;
+use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Columns\ImageColumn;
+use Filament\Tables\Filters\SelectFilter;
+use Filament\Tables\Table;
+use Firefly\FilamentBlog\Models\Post;
+use Illuminate\Support\Str;
+use Firefly\FilamentBlog\Tables\Columns\UserPhotoName;
+
+class PostsTable
+{
+    public static function configure(Table $table): Table
+    {
+        return $table
+            ->deferLoading()
+            ->columns([
+                TextColumn::make('title')
+                    ->description(function (Post $record) {
+                        return Str::limit($record->sub_title, 40);
+                    })
+                    ->searchable()
+                    ->limit(20),
+
+                TextColumn::make('status')
+                    ->badge()
+                    ->color(function ($state) {
+                        return $state->getColor();
+                    }),
+
+                ImageColumn::make('cover_photo_path')->label('Cover Photo'),
+
+                UserPhotoName::make('user')
+                    ->label('Author'),
+
+                TextColumn::make('created_at')
+                    ->dateTime()
+                    ->sortable()
+                    ->toggleable(isToggledHiddenByDefault: true),
+
+                TextColumn::make('updated_at')
+                    ->dateTime()
+                    ->sortable()
+                    ->toggleable(isToggledHiddenByDefault: true),
+            ])
+            ->filters([
+                SelectFilter::make('user')
+                    ->relationship('user', config('filamentblog.user.columns.name'))
+                    ->searchable()
+                    ->preload()
+                    ->multiple(),
+            ])
+            ->recordActions([
+                ActionGroup::make([
+                    ViewAction::make(),
+                    EditAction::make(),
+                ]),
+            ])
+            ->toolbarActions([
+                BulkActionGroup::make([
+                    DeleteBulkAction::make(),
+                ]),
+            ])
+            ->defaultSort('id', 'desc');
+    }
+}
